+++ lib/src/foundation/liquid_responsive.dart
@@ -7,8 +7,8 @@ enum LiquidScreenType { mobile, tablet, desktop }
 class LiquidBreakpoints {
   const LiquidBreakpoints._();
 
-  static const double mobileMaxWidth = 599;
-  static const double tabletMaxWidth = 1023;
+  static const double mobileMaxWidth = LiquidSizes.mobileBreakpoint - 1;
+  static const double tabletMaxWidth = LiquidSizes.tabletBreakpoint - 1;
 }
 
 class LiquidResponsive {
@@ -16,6 +16,7 @@ class LiquidResponsive {
 
   static LiquidScreenType screenTypeOf(BuildContext context) {
     final width = MediaQuery.sizeOf(context).width;
+    LiquidSizes.updateScale(width);
     if (width <= LiquidBreakpoints.mobileMaxWidth) {
       return LiquidScreenType.mobile;
     }
@@ -43,9 +44,9 @@ class LiquidResponsive {
 
   static double textScaleOf(
     BuildContext context, {
-    double mobile = LiquidResponsiveTokens.textScaleMobile,
-    double tablet = LiquidResponsiveTokens.textScaleTablet,
-    double desktop = LiquidResponsiveTokens.textScaleDesktop,
+    double mobile = LiquidTextScaleTokens.textScaleMobile,
+    double tablet = LiquidTextScaleTokens.textScaleTablet,
+    double desktop = LiquidTextScaleTokens.textScaleDesktop,
   }) {
     return value(
       context: context,
@@ -79,9 +80,9 @@ extension LiquidResponsiveContext on BuildContext {
   }
 
   double liquidTextScale({
-    double mobile = LiquidResponsiveTokens.textScaleMobile,
-    double tablet = LiquidResponsiveTokens.textScaleTablet,
-    double desktop = LiquidResponsiveTokens.textScaleDesktop,
+    double mobile = LiquidTextScaleTokens.textScaleMobile,
+    double tablet = LiquidTextScaleTokens.textScaleTablet,
+    double desktop = LiquidTextScaleTokens.textScaleDesktop,
   }) {
     return LiquidResponsive.textScaleOf(
       this,
@@ -90,22 +91,14 @@ extension LiquidResponsiveContext on BuildContext {
       desktop: desktop,
     );
   }
-
-  double liquidDouble(LiquidAdaptiveDouble value) {
-    return liquidValue<double>(
-      mobile: value.mobile,
-      tablet: value.tablet,
-      desktop: value.desktop,
-    );
-  }
 }
 
 extension LiquidResponsiveTextStyle on TextStyle {
   TextStyle liquidScale(
     BuildContext context, {
-    double mobile = LiquidResponsiveTokens.textScaleMobile,
-    double tablet = LiquidResponsiveTokens.textScaleTablet,
-    double desktop = LiquidResponsiveTokens.textScaleDesktop,
+    double mobile = LiquidTextScaleTokens.textScaleMobile,
+    double tablet = LiquidTextScaleTokens.textScaleTablet,
+    double desktop = LiquidTextScaleTokens.textScaleDesktop,
   }) {
     final baseSize = fontSize;
     if (baseSize == null) return this;