+++ lib/src/widgets/liquid_glass_dropdown.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+import '../foundation/liquid_glass_theme.dart';
 import 'liquid_glass_card.dart';
 
 class LiquidGlassDropdown<T> extends StatelessWidget {
@@ -17,6 +18,7 @@ class LiquidGlassDropdown<T> extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return LiquidGlassCard(
+      blurMode: LiquidBlurMode.real,
       borderRadius: BorderRadius.circular(14),
       padding: const EdgeInsets.symmetric(horizontal: 12),
       child: DropdownButtonHideUnderline(