+++ lib/src/widgets/liquid_glass_modal_sheet.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+import '../foundation/liquid_glass_theme.dart';
 import 'liquid_glass_card.dart';
 
 class LiquidGlassModalSheet extends StatelessWidget {
@@ -23,6 +24,7 @@ class LiquidGlassModalSheet extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return LiquidGlassCard(
+      blurMode: LiquidBlurMode.real,
       borderRadius: BorderRadius.circular(24),
       padding: const EdgeInsets.all(16),
       child: child,